Updated java code and insert statement
The insert statements are in insertstate.sql. There are 4000 users,
20,000 sets and 200,000 cards. Please use it directly to generate sql if
you find the number large enough, because I have checked the
correctness.
If not, after you rerun the java code, the new statements will be saved
in a new file call insertstatements.sql (note the extra 's'). After you
run in oracle, please use
select max(set_id) from cardSet;
to check if the max(set_id)=20000 to avoid errors in inserting cards.
(This is due to the unsafety of the increment trigger in oracle)